disease-prediction-doctors-notes/disease-prediction-doctors-notes/requirements.txt

pandas
numpy
scikit-learn
nltk
mysql-connector-python
python-dotenv
joblib
matplotlib
